mostly yambar config update, but misc edits as well

+++ .local/bin/notifications.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 
 # Notif logic
-[ -e /tmp/new_notifs ] && rm /tmp/new_notifs
-mkfifo /tmp/new_notifs
+if [ ! -p /tmp/new_notifs ]; then
+	[ -e /tmp/new_notifs ] && rm /tmp/new_notifs
+	mkfifo /tmp/new_notifs
+fi
 
 tiramisu -o "#summary\t#body\t#timeout" |
 	while read -r line; do
